remove matchable trait from testevent
