+++ src/test/spec/unified_runner/test_event.rs
@@ -1,10 +1,8 @@
 use crate::{
-    bson::{Bson, Document},
-    test::{CommandEvent, Matchable},
+    bson::Document,
+    test::CommandEvent,
 };
 use serde::Deserialize;
-
-use super::results_match;
 
 #[derive(Debug, Deserialize, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -23,85 +21,6 @@ pub enum TestEvent {
     CommandFailedEvent {
         command_name: Option<String>,
     },
-}
-
-impl Matchable for TestEvent {
-    fn content_matches(&self, actual: &TestEvent) -> bool {
-        match (self, actual) {
-            (
-                TestEvent::CommandStartedEvent {
-                    command_name: actual_command_name,
-                    database_name: actual_database_name,
-                    command: actual_command,
-                },
-                TestEvent::CommandStartedEvent {
-                    command_name: expected_command_name,
-                    database_name: expected_database_name,
-                    command: expected_command,
-                },
-            ) => {
-                if expected_command_name.is_some() && actual_command_name != expected_command_name {
-                    return false;
-                }
-                if expected_database_name.is_some()
-                    && actual_database_name != expected_database_name
-                {
-                    return false;
-                }
-                if let Some(expected_command) = expected_command {
-                    let actual_command = actual_command
-                        .as_ref()
-                        .map(|doc| Bson::Document(doc.clone()));
-                    results_match(
-                        actual_command.as_ref(),
-                        &Bson::Document(expected_command.clone()),
-                        false,
-                        None,
-                    )
-                } else {
-                    true
-                }
-            }
-            (
-                TestEvent::CommandSucceededEvent {
-                    command_name: actual_command_name,
-                    reply: actual_reply,
-                },
-                TestEvent::CommandSucceededEvent {
-                    command_name: expected_command_name,
-                    reply: expected_reply,
-                },
-            ) => {
-                if expected_command_name.is_some() && actual_command_name != expected_command_name {
-                    return false;
-                }
-                if let Some(expected_reply) = expected_reply {
-                    let actual_reply = actual_reply.as_ref().map(|doc| Bson::Document(doc.clone()));
-                    results_match(
-                        actual_reply.as_ref(),
-                        &Bson::Document(expected_reply.clone()),
-                        false,
-                        None,
-                    )
-                } else {
-                    true
-                }
-            }
-            (
-                TestEvent::CommandFailedEvent {
-                    command_name: actual_command_name,
-                },
-                TestEvent::CommandFailedEvent {
-                    command_name: expected_command_name,
-                },
-            ) => match (expected_command_name, actual_command_name) {
-                (Some(expected), Some(actual)) => expected == actual,
-                (Some(_), None) => false,
-                _ => true,
-            },
-            _ => false,
-        }
-    }
 }
 
 impl From<CommandEvent> for TestEvent {
